angle units unified to radians

+++ KheperRobotSymulation/Simulation/Entities/RectangularEnt.h
@@ -12,7 +12,7 @@ class RectangularEnt : public SimEnt
 	public:
 		// x, y -> left-bottom corner coords, rotating clockwise
 		RectangularEnt(uint16_t id, uint32_t weight, bool movable, double x,
-			double y, double width, double height, double angle = 0);
+			double y, double width, double height, float angle = 0);
 		RectangularEnt(std::ifstream& file);
 
 		virtual ~RectangularEnt() { delete _bottLeft; delete _center; }
@@ -33,7 +33,7 @@ class RectangularEnt : public SimEnt
 		Point* _center;
 		double _width;
 		double _height;
-		double _angle; // in degrees, rotating clockwise
+		float _angle; // in radians, rotating clockwise
 };
 
 #endif
